actually exporting code coverage to source file trial xml files.

+++ src/org/adligo/tests4j_tests/base_trials/I_CountingPackageTrials.java
@@ -2,15 +2,17 @@ package org.adligo.tests4j_tests.base_trials;
 
 import org.adligo.tests4j.system.shared.trials.I_MetaTrialInputData;
 
+import java.math.BigInteger;
+
 
 public interface I_CountingPackageTrials extends I_MetaTrialInputData {
 
 	public int getTrialCount();
 	
-	public int getTestCount();
+	public long getTestCount();
 	
-	public long getAssertCount();
+	public BigInteger getAssertCount();
 	
-	public long getUniqueAssertCount();
+	public BigInteger getUniqueAssertCount();
 	
 }
